Support navigation to the original element for generated source code
