+++ platform/analysis-api/src/com/intellij/openapi/roots/GeneratedSourcesFilter.java
@@ -17,7 +17,6 @@ package com.intellij.openapi.roots;
 
 import com.intellij.openapi.extensions.ExtensionPointName;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.roots.GeneratedSourcesFilterEx.LocationInFile;
 import com.intellij.openapi.util.Segment;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.psi.PsiElement;
@@ -43,50 +42,54 @@ public abstract class GeneratedSourcesFilter {
     return false;
   }
 
+  /**
+   * Collect original elements from all filters.
+   */
   @NotNull
-  public static List<? extends PsiElement> getAllOriginalElements(@NotNull PsiElement element) {
+  public static List<? extends PsiElement> collectAllOriginalElements(@NotNull PsiElement element) {
     List<PsiElement> result = null;
     for (GeneratedSourcesFilter filter : EP_NAME.getExtensions()) {
-      if (filter instanceof GeneratedSourcesFilterEx) {
-        result = addAll(((GeneratedSourcesFilterEx)filter).getOriginalElements(element), result);
-      }
+      result = addAll(filter.getOriginalElements(element), result);
     }
     return orEmpty(result);
   }
 
+  /**
+   * Collect generated elements from all filters.
+   */
   @NotNull
-  public static List<? extends PsiElement> getAllGeneratedElements(@NotNull PsiElement element) {
+  public static List<? extends PsiElement> collectAllGeneratedElements(@NotNull PsiElement element) {
     List<PsiElement> result = null;
     for (GeneratedSourcesFilter filter : EP_NAME.getExtensions()) {
-      if (filter instanceof GeneratedSourcesFilterEx) {
-        result = addAll(((GeneratedSourcesFilterEx)filter).getGeneratedElements(element), result);
-      }
+      result = addAll(filter.getGeneratedElements(element), result);
     }
     return orEmpty(result);
   }
 
+  /**
+   * Collect original locations from all filters.
+   */
   @NotNull
-  public static List<? extends LocationInFile> getAllOriginalLocations(@NotNull Project project,
-                                                                       @NotNull VirtualFile file,
-                                                                       @Nullable Segment segment) {
-    List<GeneratedSourcesFilterEx.LocationInFile> result = null;
+  public static List<? extends LocationInFile> collectAllOriginalLocations(@NotNull Project project,
+                                                                           @NotNull VirtualFile file,
+                                                                           @Nullable Segment segment) {
+    List<LocationInFile> result = null;
     for (GeneratedSourcesFilter filter : EP_NAME.getExtensions()) {
-      if (filter instanceof GeneratedSourcesFilterEx) {
-        result = addAll(((GeneratedSourcesFilterEx)filter).getOriginalLocations(project, file, segment), result);
-      }
+      result = addAll(filter.getOriginalLocations(project, file, segment), result);
     }
     return orEmpty(result);
   }
 
+  /**
+   * Collect generated locations from all filters.
+   */
   @NotNull
-  public static List<? extends LocationInFile> getAllGeneratedLocations(@NotNull Project project,
-                                                                        @NotNull VirtualFile file,
-                                                                        @Nullable Segment segment) {
-    List<GeneratedSourcesFilterEx.LocationInFile> result = null;
+  public static List<? extends LocationInFile> collectAllGeneratedLocations(@NotNull Project project,
+                                                                            @NotNull VirtualFile file,
+                                                                            @Nullable Segment segment) {
+    List<LocationInFile> result = null;
     for (GeneratedSourcesFilter filter : EP_NAME.getExtensions()) {
-      if (filter instanceof GeneratedSourcesFilterEx) {
-        result = addAll(((GeneratedSourcesFilterEx)filter).getGeneratedLocations(project, file, segment), result);
-      }
+      result = addAll(filter.getGeneratedLocations(project, file, segment), result);
     }
     return orEmpty(result);
   }
@@ -105,4 +108,99 @@ public abstract class GeneratedSourcesFilter {
   }
 
   public abstract boolean isGeneratedSource(@NotNull VirtualFile file, @NotNull Project project);
+
+  /**
+   * Returns all elements that have been processed by a code generator to derive the given element.
+   *
+   * @param element the generated element
+   * @return a list of original elements. An empty result indicates that the element is not considered to be generated by the filter.
+   */
+  @NotNull
+  public List<? extends PsiElement> getOriginalElements(@NotNull PsiElement element) {
+    return Collections.emptyList();
+  }
+
+  /**
+   * Returns all elements that have been generated from the given element.
+   *
+   * @param element the generated element
+   * @return a list of generated elements. An empty result indicates that no elements are known to be generated from the given element.
+   */
+  @NotNull
+  public List<? extends PsiElement> getGeneratedElements(@NotNull PsiElement element) {
+    return Collections.emptyList();
+  }
+
+  /**
+   * Returns all locations that have been processed by a code generator to generate the given location.
+   *
+   * @return a list of locations. An empty result indicates that the given segment is not considered to be generated by the filter.
+   */
+  @NotNull
+  public List<? extends LocationInFile> getOriginalLocations(@NotNull Project project,
+                                                             @NotNull VirtualFile file,
+                                                             @Nullable Segment segment) {
+    return Collections.emptyList();
+  }
+
+  /**
+   * Returns all locations that have been generated from the given location.
+   *
+   * @return a list of locations. An empty result indicates that no locations are known to be generated from the given segments.
+   */
+  @NotNull
+  public List<? extends LocationInFile> getGeneratedLocations(@NotNull Project project,
+                                                              @NotNull VirtualFile file,
+                                                              @Nullable Segment segment) {
+    return Collections.emptyList();
+  }
+
+  public static class LocationInFile {
+
+    @NotNull private final VirtualFile myFile;
+    @Nullable private final Segment mySegment;
+
+    public LocationInFile(@NotNull VirtualFile file, @Nullable Segment segment) {
+      myFile = file;
+      mySegment = segment;
+    }
+
+    @NotNull
+    public VirtualFile getFile() {
+      return myFile;
+    }
+
+    @Nullable
+    public Segment getSegment() {
+      return mySegment;
+    }
+
+    @Override
+    public String toString() {
+      return "LocationInFile{" +
+             "myFile=" + myFile +
+             ", mySegment=" + mySegment +
+             '}';
+    }
+
+    @Override
+    public boolean equals(Object o) {
+      if (this == o) return true;
+      if (o == null || getClass() != o.getClass()) return false;
+
+      LocationInFile that = (LocationInFile)o;
+
+      if (!myFile.equals(that.myFile)) return false;
+      if (mySegment != null ? !mySegment.equals(that.mySegment) : that.mySegment != null) return false;
+
+      return true;
+    }
+
+    @Override
+    public int hashCode() {
+      int result = myFile.hashCode();
+      result = 31 * result + (mySegment != null ? mySegment.hashCode() : 0);
+      return result;
+    }
+  }
 }
